+++ packages/gestalt/src/Contents.js
@@ -4,6 +4,7 @@ import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import Caret from './Caret.js';
 import styles from './Contents.css';
+import borders from './Borders.css';
 import colors from './Colors.css';
 
 /* Needed until this Flow issue is fixed: https://github.com/facebook/flow/issues/380 */
@@ -54,6 +55,7 @@ type EdgeShift = { caret: Shift, flyout: Shift };
 
 type Props = {|
   bgColor: 'blue' | 'darkGray' | 'orange' | 'red' | 'white',
+  border?: boolean,
   caret?: boolean,
   children?: React.Node,
   idealDirection?: 'up' | 'right' | 'down' | 'left',
@@ -64,6 +66,7 @@ type Props = {|
     x: number,
     y: number,
   },
+  rounding?: 2 | 4,
   shouldFocus?: boolean,
   triggerRect: ClientRect,
   width: ?number,
@@ -331,6 +334,7 @@ export function baseOffsets(
 export default class Contents extends React.Component<Props, State> {
   static propTypes = {
     bgColor: PropTypes.oneOf(['blue', 'darkGray', 'orange', 'red', 'white']),
+    border: PropTypes.bool,
     caret: PropTypes.bool,
     children: PropTypes.node,
     idealDirection: PropTypes.oneOf(['up', 'right', 'down', 'left']),
@@ -341,6 +345,7 @@ export default class Contents extends React.Component<Props, State> {
       y: PropTypes.number,
     }),
     positionRelativeToAnchor: PropTypes.bool,
+    rounding: PropTypes.oneOf([2, 4]),
     shouldFocus: PropTypes.bool,
     triggerRect: PropTypes.exact({
       bottom: PropTypes.number,
@@ -354,6 +359,7 @@ export default class Contents extends React.Component<Props, State> {
   };
 
   static defaultProps = {
+    border: true,
     caret: true,
   };
 
@@ -485,7 +491,7 @@ export default class Contents extends React.Component<Props, State> {
   };
 
   render() {
-    const { bgColor, caret, children, width } = this.props;
+    const { bgColor, border, caret, children, rounding, width } = this.props;
     const { caretOffset, flyoutOffset, mainDir } = this.state;
 
     // Needed to prevent UI thrashing
@@ -500,8 +506,11 @@ export default class Contents extends React.Component<Props, State> {
       >
         <div
           className={classnames(
+            border && styles.border,
             colors[background],
             colors[bgColor],
+            rounding === 2 && borders.rounding2,
+            rounding === 4 && borders.rounding4,
             styles.contents,
             styles.maxDimensions,
             width !== null && styles.minDimensions