+++ packages/gestalt/src/Tooltip.jsdom.test.js
@@ -22,7 +22,7 @@ test('Tooltip renders the link when hovered', () => {
       link={
         <Link href="https://pinterest.com" target="blank">
           <Text color="white" size="sm" weight="bold">
-            Learn more about logout
+            Learn more
           </Text>
         </Link>
       }
@@ -34,7 +34,7 @@ test('Tooltip renders the link when hovered', () => {
 
   fireEvent.mouseEnter(container.querySelector('[aria-label]'));
   const { body } = document;
-  const element = getByText('Learn more about logout');
+  const element = getByText('Learn more');
   expect(body && body.contains(element)).toBeTruthy();
 });